Experimenting to try and find a way to get emojis to be stored on rds

+++ play-java-jpa-example/app/models/Person.java
@@ -139,7 +139,7 @@ public class Person {
         Collections.sort(postList, new Comparator<Post>() {
             @Override
             public int compare(Post o1, Post o2) {
-                return Math.toIntExact(o2.timeStamp - o1.timeStamp);
+                return (int) (o2.timeStamp - o1.timeStamp);
             }
         });
